+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/assets/articulation/articulation.py
@@ -516,6 +516,8 @@ class Articulation(RigidObject):
         # process configuration
         self._process_cfg()
         self._process_actuators_cfg()
+        # validate configuration
+        self._validate_cfg()
         # log joint information
         self._log_articulation_joint_info()
 
@@ -698,6 +700,46 @@ class Articulation(RigidObject):
     """
     Internal helpers -- Debugging.
     """
+
+    def _validate_cfg(self):
+        """Validate the configuration after processing.
+
+        Note:
+            This function should be called only after the configuration has been processed and the buffers have been
+            created. Otherwise, some settings that are altered during processing may not be validated.
+            For instance, the actuator models may change the joint max velocity limits.
+        """
+        # check that the default values are within the limits
+        joint_pos_limits = self.root_physx_view.get_dof_limits()[0].to(self.device)
+        out_of_range = self._data.default_joint_pos[0] < joint_pos_limits[:, 0]
+        out_of_range |= self._data.default_joint_pos[0] > joint_pos_limits[:, 1]
+        violated_indices = torch.nonzero(out_of_range, as_tuple=False).squeeze(-1)
+        # throw error if any of the default joint positions are out of the limits
+        if len(violated_indices) > 0:
+            # prepare message for violated joints
+            msg = "The following joints have default positions out of the limits: \n"
+            for idx in violated_indices:
+                joint_name = self.data.joint_names[idx]
+                joint_limits = joint_pos_limits[idx]
+                joint_pos = self.data.default_joint_pos[0, idx]
+                # add to message
+                msg += f"\t- '{joint_name}': {joint_pos:.3f} not in [{joint_limits[0]:.3f}, {joint_limits[1]:.3f}]\n"
+            raise ValueError(msg)
+
+        # check that the default joint velocities are within the limits
+        joint_max_vel = self.root_physx_view.get_dof_max_velocities()[0].to(self.device)
+        out_of_range = torch.abs(self._data.default_joint_vel[0]) > joint_max_vel
+        violated_indices = torch.nonzero(out_of_range, as_tuple=False).squeeze(-1)
+        if len(violated_indices) > 0:
+            # prepare message for violated joints
+            msg = "The following joints have default velocities out of the limits: \n"
+            for idx in violated_indices:
+                joint_name = self.data.joint_names[idx]
+                joint_limits = [-joint_max_vel[idx], joint_max_vel[idx]]
+                joint_vel = self.data.default_joint_vel[0, idx]
+                # add to message
+                msg += f"\t- '{joint_name}': {joint_vel:.3f} not in [{joint_limits[0]:.3f}, {joint_limits[1]:.3f}]\n"
+            raise ValueError(msg)
 
     def _log_articulation_joint_info(self):
         """Log information about the articulation's simulated joints."""